Fixing up CSS files and the border around the map
