+++ frontend/my-vite-app/src/Map.jsx
@@ -1,31 +1,37 @@
-import React, { useEffect, useRef } from 'react';
-import mapboxgl from 'mapbox-gl';
+import React from 'react';
+import { Map, NavigationControl } from 'react-map-gl/mapbox';
 import 'mapbox-gl/dist/mapbox-gl.css';
 
-const MAPBOX_TOKEN = 'pk.eyJ1IjoidmRhbGZhcm8iLCJhIjoiY21iMDk2MnA3MG9sYzJrcHNveXJ2MnQ2cyJ9.nSBsNvmgeK-6kyHM2-9h2g';
+const TOKEN = 'pk.eyJ1IjoidmRhbGZhcm8iLCJhIjoiY21iMDk2MnA3MG9sYzJrcHNveXJ2MnQ2cyJ9.nSBsNvmgeK-6kyHM2-9h2g';
 
-mapboxgl.accessToken = MAPBOX_TOKEN;
-
-const Map = () => {
-  const mapContainer = useRef(null);
-  const map = useRef(null);
-
-  useEffect(() => {
-    if (map.current) return; // initialize map only once
-
-    map.current = new mapboxgl.Map({
-      container: mapContainer.current,
-      style: 'mapbox://styles/mapbox/streets-v11',
-      center: [-122.4376, 37.7577],
-      zoom: 10.5,
-    });
-
-    map.current.addControl(new mapboxgl.NavigationControl(), 'top-right');
-
-    return () => map.current.remove();
-  }, []);
-
-  return <div ref={mapContainer} style={{ width: '100%', height: '100vh' }} />;
+const initialView = {
+  longitude: -122.43,
+  latitude: 37.78,
+  zoom: 10.5,
 };
 
-export default Map;
+export default function MapComponent() {
+  return (
+    <div
+      style={{
+        position: 'relative',
+        height: '500px',
+        width: '1000px',
+        display: 'block', // ✅ Ensures it's treated as a block element
+      }}
+    >
+      <Map
+        initialViewState={initialView}
+        mapStyle="mapbox://styles/mapbox/light-v9"
+        mapboxAccessToken={TOKEN}
+        style={{
+          width: '100%',
+          height: '100%',
+          display: 'block', // ✅ Prevents collapsing map
+        }}
+      >
+        <NavigationControl style={{ position: 'absolute', top: 10, right: 10 }} />
+      </Map>
+    </div>
+  );
+}
